// src/app/common/models/interface/jugador.interface.ts
export interface IJugador {
  id: string;
  nombre: string;
  telefono: string;
  sexo: string;
  foto: string;
  correo: string;
  nro_identidad: string;
  alias: string;
  fecha_nacimiento: string;
  ocupacion: string;
  comentario: string;
  nivel: string;
  elo: string;
  ranking: string;
  tipo: string;
  ciudad_id: string;
}